rename static-site/src/components/sars-cov-2;
this directory contains components
which are now generally applicable
to any page that wants to display a
collection of info about any grouping
of builds (e.g. flu, groups, commeunity),
not just for sars-cov-2.
