+++ static-site/src/components/build-pages/sit-reps.jsx
